Added login test and tests to pipeline
