+++ config.toml
@@ -1,4 +1,4 @@
-CFN = "LookAtMySweetCan"
+CFN = "GreenSoap"
 
 # Example
 # CFN = "GreenSoap"
