+++ errors.go
@@ -1,21 +1,31 @@
 package main
 
 type AppError struct {
-	message string
+	message    string
+	returnCode int
 }
 
 var LoginError = AppError{
-	message: `Failed to Log in`,
+	message:    `Failed to Log in`,
+	returnCode: -1,
 }
 
 var ProfileError = AppError{
-	message: `CFN profile does not exist`,
+	message:    `CFN profile does not exist`,
+	returnCode: -2,
+}
+
+var CaptchaError = AppError{
+	message:    `Encountered a captcha, please wait a while before opening the application again.`,
+	returnCode: -3,
 }
 
 var ParseError = AppError{
-	message: `Could not locate data on CFN profile`,
+	message:    `Could not locate data on CFN profile`,
+	returnCode: -4,
 }
 
 var SaveErrror = AppError{
-	message: `Could not save data`,
+	message:    `Could not save data`,
+	returnCode: -5,
 }
